changes to remove library functionality out of core and allow for library to use cement config parsing
