+++ bionetgen/__init__.py
@@ -1,4 +1,4 @@
-from .core.main import run
 from .core.defaults import defaults
 from .xmlapi import bngmodel
+from .xmlapi.runner import run
 from .simulator import sim_getter
