+++ bionetgen/core/__init__.py
@@ -1,2 +1,5 @@
+# NOTE Anything that needs to go into the library
+# side needs to not be in the core section, it 
+# leads to circular imports
 from .result import BNGResult
 from .plot import BNGPlotter
